Crate cargo_crev_web_admin

cargo_crev_web_admin lib.rs

cargo_crev_web_admin

Admin CLI for cargo_crev_web
repository; version: 2022.126.1230 date: 2022-01-26 authors: Luciano Bestia

cargo_crev_web_admin CLI

The web app cargo_crev_web on https://web.crev.dev fetches all proof repos it can find and shows the crate reviews online.
Some admin tasks are needed and I don’t want them to be accessible on the web.
This will be a CLI app that can be used when logged on the linux terminal over SSH.
So is sure that only an admin, who can log in on to the server, can use this tasks.

delete from trust someone from blocklist json (case insensitive)

Delete from \\Secure FTP\google_cloud\home\luciano_bestia.cache\crev\remotes
folders of reviewers that are not in \\Secure FTP\google_cloud\home\luciano_bestia\config\crev\proofs\github_com_cargo-crev-web_crev-proofs-.....\trust

short command for cargo crev id query trusted –high-cost 1 –medium-cost 1 –low-cost 1 –depth 1

short command for cargo crev repo fetch trusted –high-cost 1 –medium-cost 1 –low-cost 1 –depth 1

short command for web reindex

short command for cargo crev trust –level low https://github.com/Alxandr/crev-proofs

command to add to blocklist

command to delete from blocklist

Integrity - warnings if a review have incorrect url or ID

Development

I use cargo-auto for automation tasks in rust language. Install it:

cargo install cargo-auto

List user-defined automation tasks in automation_tasks_rs:

cargo auto

bash auto-completion

This executable is prepared for auto-completion in bash.
Run this command to define auto-completion in bash for the current session:

complete -C "cargo_crev_web_admin completion" cargo_crev_web_admin

To make it permanent add this command to the file ~/.bashrc or some other file that runs commands on bash initialization.

TODO

all

cargo crev reviews and advisory

Please, spread this info !
Open source code needs a community effort to express trustworthiness.
Start with reading the reviews of the crates on web.crev.dev.
Then install the GUI cargo_crev_reviews or the CLI cargo-crev.
Your personal reviews are most important. If you have a boss, he will sooner or later ask you to show him your reviews for all the dependencies you use. With cargo_crev_reviews you have a basic tool to do that.
Write your reviews! Describe the crates you trust and why. Or warn about the crate versions you think are dangerous. Publish and share your opinion with other developers.

open-source and free as a beer

My open-source projects are free as a beer (MIT license).
I just love programming.
But I need also to drink. If you find my projects and tutorials helpful,
please buy me a beer or two donating on my paypal.
You know the price of a beer in your local bar ;-)
So I can drink a free beer for your health :-)
Na zdravje !

Modules

utils_mod

A module with often used functions.

Structs

CLEAR_ALL

ansi clear all

CLEAR_LINE

ansi clear line

GREEN

ansi color

RED

ansi color

RESET

ansi reset color

YELLOW

ansi color

Functions

ns_elapsed

returns the elapsed nanoseconds

ns_print_ms

print elapsed time in milliseconds and returns the new now in nanoseconds

ns_print_ns

print elapsed time in nanoseconds and returns the new now in nanoseconds

ns_start

returns the now in nanoseconds

trusted

list the explicit trusted reviewers from the /trust/*.crev files