Crate cargo_crev_web_admin

cargo_crev_web_admin

Admin CLI for cargo_crev_web
repository; version: 2022.128.1153 date: 2022-01-28 authors: Luciano Bestia

cargo_crev_web_admin CLI

The web app cargo_crev_web on https://web.crev.dev fetches all proof repos it can find and shows the crate reviews online.
Some admin tasks are needed and I don’t want them to be accessible on the web.
This will be a CLI app that can be used when logged on the linux terminal over SSH.
So is sure that only an admin, who can log in on to the server, can use this tasks.

Some tasks need the crev passphrase. Put it in the env variable before starting the CLI:
$ export CREV_PASSPHRASE=xxx
Add a space before the command to avoid to be saved in the bash history.

Development

I use cargo-auto for automation tasks in rust language. Install it:

cargo install cargo-auto

List user-defined automation tasks in automation_tasks_rs:

cargo auto

bash auto-completion

This executable is prepared for auto-completion in bash.
Run this command to define auto-completion in bash for the current session:
Or add it to .bashrc file to be executed n every session start.

complete -C "cargo_crev_web_admin completion" cargo_crev_web_admin

To make it permanent add this command to the file ~/.bashrc or some other file that runs commands on bash initialization.

TODO

Integrity - warnings if a review have incorrect url or ID

cargo crev reviews and advisory

Please, spread this info !
Open source code needs a community effort to express trustworthiness.
Start with reading the reviews of the crates on web.crev.dev.
Then install the GUI cargo_crev_reviews or the CLI cargo-crev.
Your personal reviews are most important. If you have a boss, he will sooner or later ask you to show him your reviews for all the dependencies you use. With cargo_crev_reviews you have a basic tool to do that.
Write your reviews! Describe the crates you trust and why. Or warn about the crate versions you think are dangerous. Publish and share your opinion with other developers.

open-source and free as a beer

My open-source projects are free as a beer (MIT license).
I just love programming.
But I need also to drink. If you find my projects and tutorials helpful,
please buy me a beer donating on my paypal.
You know the price of a beer in your local bar ;-)
So I can drink a free beer for your health :-)
Na zdravje! Alla salute! Prost! Nazdravlje! 🍻

Modules

blocklisted_repos_mod

It is only one json file. Serialize and deserialize, read and write, add and delete

my_trusted_repos_mod

utils_mod

A module with often used functions.

Structs

CLEAR_ALL

ansi clear all

CLEAR_LINE

ansi clear line

CREV_REMOTES_DIR

GREEN

ansi color

HOME_DIR

RED

ansi color

RESET

ansi reset color

YELLOW

ansi color

Functions

blocklisted_add

add new blocklist repo

blocklisted_delete

delete from blocklist repo

blocklisted_list

list of blocklisted

count_newlines

datetime_now_for_file_names

fetch

fetch the explicit trusted reviewers from the /trust/*.crev files

ns_elapsed

returns the elapsed nanoseconds

ns_print_ms

print elapsed time in milliseconds and returns the new now in nanoseconds

ns_print_ns

print elapsed time in nanoseconds and returns the new now in nanoseconds

ns_start

returns the now in nanoseconds

publish_to_github

after changing trust files it is mandatory to publish this repo

reindex

web app reads and reindex new or changed data

remotes_delete

delete fetched repos from /remote/ if they are not in trusted_list

trusted_add

add new trusted repo

trusted_delete

delete from trusted repo

trusted_from_crev_command

list the explicit trusted reviewers from cargo crev command

trusted_list

list the explicit trusted reviewers from the /trust/*.crev files